move MessageProperties type to types-live and getMessageProperties to coin-evm/logic
Remove duplication of MessageProperties type definition and
getMessageProperties function definition between LLD and LLM
Since only evm currencies seem to use getMessageProperties as of today,
it has been moved to coin-evm
