+++ apps/ledger-live-desktop/src/renderer/families/evm/index.ts
@@ -1,13 +1,13 @@
-import { EvmFamily } from "./types";
-import { injectNftIntoTransaction, getNftTransactionProperties } from "./nft";
+import { getMessageProperties } from "@ledgerhq/coin-evm/logic";
 import AccountFooter from "./AccountFooter";
-import operationDetails from "./operationDetails";
 import accountHeaderManageActions from "./AccountHeaderManageActions";
-import transactionConfirmFields from "./TransactionConfirmFields";
 import sendAmountFields from "./SendAmountFields";
-import StepSummaryNetworkFeesRow from "./StepSummaryNetworkFeesRow";
 import StakeBanner from "./StakeBanner";
-import { getMessageProperties } from "./message";
+import StepSummaryNetworkFeesRow from "./StepSummaryNetworkFeesRow";
+import transactionConfirmFields from "./TransactionConfirmFields";
+import { getNftTransactionProperties, injectNftIntoTransaction } from "./nft";
+import operationDetails from "./operationDetails";
+import { EvmFamily } from "./types";
 
 const family: EvmFamily = {
   AccountFooter,
