+++ libs/ledgerjs/packages/types-live/src/messages/index.ts
@@ -7,4 +7,9 @@ export type DefaultMessage = {
 
 export type AnyMessage = DefaultMessage | TypedEvmMessage;
 
+export type MessageProperties = {
+  label: string;
+  value: string | string[];
+}[];
+
 export * from "./evm";
